requirements.txt
feedparser>=6.0.10
httpx>=0.27.0
PyYAML>=6.0
python-dateutil>=2.8.2
PyGithub>=2.1.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-pptx>=0.6.23
markdown-it-py>=3.0.0
pytest>=8.0.0
